replaced usepagination to utils/paginationutils and renamed to getpagination; added a separate type for paginationbarprops and pagilnksprops; changed arrow image to span; removed spaces from index.tsx
